Finalize quarto translation

+++ tools/render_quarto.py
@@ -14,30 +14,34 @@ from tools.texquarto.toc import TOC
 from tools.texquarto.util import get_template
 from tools.readbbl import read_bbl
 
+
 logging.basicConfig(level=logging.INFO, format='[%(asctime)s %(name)-12s %(levelname)-5s] %(message)s')
 
 parser = argparse.ArgumentParser(description='Convert CCS book latex into Quarto')
+parser.add_argument('outfolder', type=Path)
 parser.add_argument('chapters', type=int, nargs='*')
 args = parser.parse_args()
 
 bibliography = read_bbl("main.bbl")
 
 base = Path.cwd()
-out = base / "quarto"
+out = args.outfolder
 out.mkdir(exist_ok=True)
 toc = TOC(base)
 template = get_template('_quarto.yml')
 
 if args.chapters:
-    keep = set(args.chapters) | {int(x.name.replace("chapter", "").replace(".qmd", "")) for x in out.glob("chapter*.qmd")}
+    keep = set(args.chapters) | {int(x.name.replace("chapter", "").replace(".qmd", "")) for x in out.glob("chapter*.qmd") if "-" not in x.name}
     chapters = [x for x in toc.chapters if (not args.chapters) or (int(x.nr) in keep)]
 else:
     chapters = toc.chapters
 
 (out / "_quarto.yml").open("w").write(template.render(**locals()))
 
-shutil.copy((Path(template.filename).parent) / "index.qmd", out/"index.qmd")
-shutil.copy("references.bib", out/"references.bib")
+if not (out/"index.qmd").exists():
+    shutil.copy((Path(template.filename).parent) / "index.qmd", out/"index.qmd")
+if not (out/"references.bib").exists():
+    shutil.copy("references.bib", out/"references.bib")
 shutil.copy((Path(template.filename).parent) / "references.qmd", out/"references.qmd")
 
 template = get_template('chapter.qmd')
@@ -50,8 +54,8 @@ for chapnr, chapter in enumerate(toc.chapters, start=1):
     outf = out / chapter.fn
     print(f"{chapter.nr}: {chapter.texfile} -> {outf}")
     current_chapter = chapter.fn
-    current_chapter_py = None if chapnr == 1 else f"chapter{chapnr:02}/chapter_{chapnr:02}_py.ipynb"
-    current_chapter_r = None if chapnr == 1 else f"chapter{chapnr:02}/chapter_{chapnr:02}_r.ipynb"
+    current_chapter_py = None if chapnr in (1,16) else f"chapter{chapnr:02}/chapter_{chapnr:02}_py.ipynb"
+    current_chapter_r = None if chapnr in (1,16) else f"chapter{chapnr:02}/chapter_{chapnr:02}_r.ipynb"
     with open(chapter.texfile) as source:
         tex, verbs = preprocess(source.read())
 
@@ -66,16 +70,3 @@ for chapnr, chapter in enumerate(toc.chapters, start=1):
     #if parser.unknown_nodes:
     #    unknown[chapter.nr] = parser.unknown_nodes
 
-sys.exit()
-print(f"** Index")
-current_chapter = "index"
-inf = get_template("index.html")
-outf = out / "index.html"
-current_chapter_py = current_chapter_r = None
-html = inf.render(**locals())
-
-outf.open("w").write(html)
-
-if unknown:
-    for chapter, missing in unknown.items():
-        logging.warning(f"Chapter {chapter} Unknown nodes: {missing}")
